Create Multithreaded code to multiply 3 matrices of higher order (upto 10^7)
Write a Java program that reads input from a file (the filename is passed as a commandline input) and writes the final output to the console. The input provided will contain 3 nXn matrices (A, B and C) in a format described below. The program should compute the product (multiplication) of the 3 input matrices ((AxB)xC), and print the output to the console according to the format described below. Your program should employ multiple Java threads for the computation, as execution time will be an evaluation parameter. 
In the program, you need to wait for all the threads to complete execution and only then exit the program. You will also need to add all the relevant exception-handling logic based on your implementation (read below on guidelines for exception handling).
Please note that how many threads to create and how to use them within the program is completely up to your implementation.

Constraints:
1 <= n <= 4000
1 <= Matrix[i][j] <= 1000000