Add files via upload
Write a Java program that reads one integer, n, from the console. The program should give the output of the sum of all the prime numbers between 1 and n (both inclusive); and the sum of all the multiples of 3 or 5 or 7 between 1 and n (both inclusive). Your program should employ multiple threads for the computation. The program should print the output to the console according to the format below.

In the program, you have to wait for all the threads to complete execution and only then exit the program. You will also need to add all the relevant exception-handling logic based on your implementation, choosing from the glossary of exceptions at the end of this document.

Please note that how many threads and how to use them within the program is completely up to your implementation.

Constraints:
 10 <= n <= 500000000