Add files via upload
Write a Java program that sorts an array of integers using multiple threads.
The program divides the sorting task among multiple threads and performs parallel sorting to improve the performance of sorting large arrays. You need to use merge sort as the sorting algorithm, and you will have to implement it from scratch. The input will be read from a file, the filename is supplied as a commandline argument. The file contains space separated integers.

In the program, you have to wait for all the threads to complete execution and then only exit the program. You will also need to add all the relevant exception-handling logic based on your implementation.

Please note that how many threads and how to use them within the program is completely up to your implementation.

Constraints:
1 <= number of integers in the file <= 1000
0 <= Array[i] <= 1000000